requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
slack_sdk==3.33.5
flask==3.1.0
python-dotenv==1.0.1